Refactor dashboard code into its own file
- To help clear out `server.js` and make more like the Sensors repo
