+++ server/chatbotDashboard.mst
@@ -175,7 +175,7 @@
             </li>
             <li class="nav-item">
                 <button class="btn btn-navbar" id="export-csv">
-                    <a href="/buttons-data" class="text-decoration-none">Export CSV</a>
+                    <a href="/export-data" class="text-decoration-none">Export CSV</a>
                 </button>
             </li>
             <li class="nav-item">
